Fix malformed closing tag
When closing a tag (block, or other), the search function did not
consider that the start and end tags can overlap in their symbols. It
also assumes a minimum length for the matching sequence based on a
complete start and complete end tag. These combined would cause range
errors by accessing the incorrect content such as:

	{#}

	// src/compiler.rs:58
	comment = tag[2..1].trim()

This is fixed by skipping the start tag for the purpose of searching for
the end tag, assuming the start tag has equal with.